add all changes to userdashboard
